requirements.txt
crewai==0.95.0
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
Flask_SocketIO==5.5.1
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
langchain==0.3.14
llama_index==0.12.11
openai==1.59.8
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
simple_salesforce==1.12.6
twilio==9.4.3
WTForms==3.2.1
